+++ proto/umee/incentive/v1/genesis.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package umeenetwork.umee.incentive.v1;
 
-option go_package = "github.com/umee-network/umee/v3/x/incentive/types";
+option go_package = "github.com/umee-network/umee/v3/x/incentive";
 
 import "cosmos/base/v1beta1/coin.proto";
 import "gogoproto/gogo.proto";
