+++ proto/umee/incentive/v1/incentive.proto
@@ -4,7 +4,7 @@ package umeenetwork.umee.incentive.v1;
 import "cosmos/base/v1beta1/coin.proto";
 import "gogoproto/gogo.proto";
 
-option go_package = "github.com/umee-network/umee/v3/x/incentive/types";
+option go_package = "github.com/umee-network/umee/v3/x/incentive";
 
 // Params defines the parameters for the incentive module.
 message Params {
