+++ proto/umee/incentive/v1/tx.proto
@@ -7,7 +7,7 @@ import "cosmos/msg/v1/msg.proto";
 import "gogoproto/gogo.proto";
 import "umee/incentive/v1/incentive.proto";
 
-option go_package = "github.com/umee-network/umee/v3/x/incentive/types";
+option go_package = "github.com/umee-network/umee/v3/x/incentive";
 
 // Msg defines the x/incentive module's Msg service.
 service Msg {
