+++ x/incentive/errors.go
@@ -1,4 +1,4 @@
-package types
+package incentive
 
 // DONTCOVER
